Have opsimulate deploy command generate SSH key and fetch Gitlab Terraform repo

+++ opsimulate/cli.py
@@ -2,10 +2,12 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2017 James Wen
 
+from Crypto.PublicKey import RSA
 import os
 import zipfile
 
 import click
+import git
 from requests import get
 
 
@@ -18,7 +20,10 @@ def cli():
 def deploy():
     # Set max charge/budget constrictions on student’s GCP account
     # Use Terraform and Gitlab terraform template to deploy and setup
+    _generate_ssh_key()
     _setup_terraform()
+    _get_gitlab_terraform()
+    # _run_terraform()
     # Setup the student’s investigative tools on the server: ensure common
     # operational tools like curl, iostat, lsop, lsof, w, dig, nslookup, mtr,
     # etc. are present
@@ -28,24 +33,54 @@ def deploy():
     pass
 
 
+def _generate_ssh_key():
+    DIR_NAME = "./keys"
+    if not os.path.isdir(DIR_NAME):
+        print("Generating SSH key")
+        os.mkdir(DIR_NAME)
+
+        key = RSA.generate(2048)
+        with open("./keys/private.key", 'w') as content_file:
+            os.chmod("./keys/private.key", 0600)
+            content_file.write(key.exportKey('PEM'))
+        pubkey = key.publickey()
+        with open("./keys/public.key", 'w') as content_file:
+            content_file.write(pubkey.exportKey('OpenSSH'))
+        print("Generated SSH key")
+
+
 def _setup_terraform():
     """
     Setup the dependencies necessary to deploy the simulation software
     """
-    # Download terraform zip
-    print("Downloading Terraform now...")
-    terraform_download_link = 'https://releases.hashicorp.com/terraform/' \
-                              '0.9.10/terraform_0.9.10_darwin_amd64.zip'
-    terraform_zip_file = 'terraform.zip'
-    with open(terraform_zip_file, "wb") as file:
-        response = get(terraform_download_link)
-        file.write(response.content)
+    if not os.path.isfile('terraform'):
+        # Download terraform zip
+        print("Downloading Terraform now...")
+        terraform_download_link = 'https://releases.hashicorp.com/terraform/' \
+                                  '0.9.10/terraform_0.9.10_darwin_amd64.zip'
+        terraform_zip_file = 'terraform.zip'
+        with open(terraform_zip_file, "wb") as file:
+            response = get(terraform_download_link)
+            file.write(response.content)
 
-    # unzip
-    zip_ref = zipfile.ZipFile(terraform_zip_file, 'r')
-    zip_ref.extractall('.')
-    zip_ref.close()
-    # terraform binary is available
-    print("Terraform is ready")
+        # unzip
+        zip_ref = zipfile.ZipFile(terraform_zip_file, 'r')
+        zip_ref.extractall('.')
+        zip_ref.close()
+        # terraform binary is available
+        print("Terraform is ready")
 
-    os.remove(terraform_zip_file)
+        os.remove(terraform_zip_file)
+
+
+def _get_gitlab_terraform():
+    DIR_NAME = "/tmp/gitlab-terraform"
+    if not os.path.isdir(DIR_NAME):
+        print("Fetching gitlab GCE terraform repo")
+        REMOTE_URL = "https://gitlab.com/gitlab-terraform/gce.git"
+        os.mkdir(DIR_NAME)
+        repo = git.Repo.init(DIR_NAME)
+        origin = repo.create_remote('origin', REMOTE_URL)
+        origin.fetch()
+        origin.pull(origin.refs[0].remote_head)
+        print("Fetched gitlab GCE terraform repo")
